спитч

+++ src/main/java/ru/plot/repo/BankRepository.java
@@ -3,5 +3,7 @@ package ru.plot.repo;
 import org.springframework.data.repository.CrudRepository;
 import ru.plot.entity.Bank;
 
-public interface BankRepository extends CrudRepository<Bank, Long> {
+import java.math.BigInteger;
+
+public interface BankRepository extends CrudRepository<Bank, BigInteger> {
 }
