+++ src/main/java/ru/plot/repo/OrganizationsRepository.java
@@ -3,5 +3,7 @@ package ru.plot.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.plot.entity.Organizations;
 
-public interface OrganizationsRepository extends JpaRepository<Organizations, Long> {
+import java.math.BigInteger;
+
+public interface OrganizationsRepository extends JpaRepository<Organizations, BigInteger> {
 }
